+++ com.oracle.truffle.r.test.native/packages/testrffi/testrffi/R/testrffi.R
@@ -1,4 +1,4 @@
-# Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2019, 2020, Oracle and/or its affiliates. All rights reserved.
 # DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
 #
 # This code is free software; you can redistribute it and/or modify it
@@ -335,3 +335,12 @@ rffi.testMissingArgWithATTRIB <- function() {
 rffi.testPRIMFUN <- function(fun, args) {
     .Call('testPRIMFUN', fun, args)
 }
+
+rffi.test_trace <- function(a, b) {
+    .Call("testTrace")
+}
+
+rffi.div_zero <- function() {
+    .Call("testdiv", 0)
+}
+